requirements.txt
BeautifulSoup
requests
python-dotenv
